refactor(workflow-tasks): move images component into library
